+++ src/Apha.BST/Apha.BST.Web/Controllers/TrainingController.cs
@@ -1,5 +1,8 @@
-﻿using Apha.BST.Application.DTOs;
+﻿using System.Runtime.CompilerServices;
+using Apha.BST.Application.DTOs;
 using Apha.BST.Application.Interfaces;
+using Apha.BST.Application.Services;
+using Apha.BST.Core.Entities;
 using Apha.BST.Core.Interfaces;
 using Apha.BST.Web.Models;
 using AutoMapper;
@@ -11,17 +14,17 @@ namespace Apha.BST.Web.Controllers
     public class TrainingController : Controller
     {
         private readonly ITrainingService _trainingService;
-        private readonly IPersonsRepository _personRepository;
+        private readonly IPersonsService _personService;
         private readonly IMapper _mapper;
-        private const string siteName = "All";
-        private readonly ILogger<TrainingController> _logger;
+        private readonly IStaticDropdownService _staticDropdownService;         
+        private const string traineeAll = "All";
 
-        public TrainingController(ITrainingService trainingService, IMapper mapper, IPersonsRepository personRepository, ILogger<TrainingController> logger)
+        public TrainingController(ITrainingService trainingService, IMapper mapper, IPersonsService personService,IStaticDropdownService staticDropdownService)
         {
             _trainingService = trainingService;
             _mapper = mapper;
-            _personRepository = personRepository;
-            _logger = logger;
+            _personService = personService;
+            _staticDropdownService = staticDropdownService;
         }
         public IActionResult Index()
         {
@@ -37,6 +40,14 @@ namespace Apha.BST.Web.Controllers
                     .ToList()
             };
 
+            viewModel.TrainingTypesList = _staticDropdownService.GetTrainingTypes()
+                 .Select(t => new SelectListItem { Value = t.Value, Text = t.Text })
+                 .ToList();
+
+            viewModel.TrainingAnimalList = _staticDropdownService.GetTrainingAnimal()
+               .Select(t => new SelectListItem { Value = t.Value, Text = t.Text })
+               .ToList();
+            viewModel.TrainingDateTime = DateTime.Today; // Set default date to now
             return View(viewModel);
         }
 
@@ -49,12 +60,19 @@ namespace Apha.BST.Web.Controllers
                 viewModel.Persons = (await _trainingService.GetTraineesAsync())
                     .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person })
                     .ToList();
+                viewModel.TrainingTypesList = _staticDropdownService.GetTrainingTypes()
+                 .Select(t => new SelectListItem { Value = t.Value, Text = t.Text })
+                 .ToList();
+
+                viewModel.TrainingAnimalList = _staticDropdownService.GetTrainingAnimal()
+                   .Select(t => new SelectListItem { Value = t.Value, Text = t.Text })
+                   .ToList();
+
                 return View(viewModel);
             }
             try
-            {
-                viewModel.TrainingDateTime = DateTime.Now;
-                var dto = _mapper.Map<TrainingDTO>(viewModel);
+            {               
+                var dto = _mapper.Map<TrainingDto>(viewModel);
                 var message = await _trainingService.AddTrainingAsync(dto);
                 TempData["Message"] = message;
             }
@@ -67,20 +85,41 @@ namespace Apha.BST.Web.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> EditTraining(int trainerId, string species, DateTime dateTrained)
+        public async Task<IActionResult> EditTraining(int traineeId, int trainerId, string species, DateTime dateTrained, string trainingType)
         {
-            var dto = await _trainingService.GetTrainingByKeysAsync(trainerId, species, dateTrained);
 
+            var dto = await _trainingService.GetTrainingByKeysAsync(traineeId, trainerId, species, dateTrained, trainingType);
             if (dto == null)
             {
-                return NotFound();
+                TempData["Message"] = "Invalid data provided for deletion.";
+                return RedirectToAction(nameof(ViewTraining));
             }
-
-            var viewModel = _mapper.Map<EditTrainingViewModel>(dto);
-            viewModel.Persons = (await _trainingService.GetTraineesAsync())
-                .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person })
+            var persons = await _trainingService.GetTraineesAsync();
+            var viewModel = new EditTrainingViewModel
+            {
+                TraineeId = dto.PersonId,
+                TrainerId = dto.TrainerId,
+                TrainingType = dto.TrainingType,
+                TrainingAnimal = dto.TrainingAnimal,
+                TrainingDateTime = dto.TrainingDateTime,
+                TrainingDateTimeOld = dto.TrainingDateTime,
+                TraineeIdOld = dto.PersonId,
+                TrainerIdOld = dto.TrainerId,
+                TrainingAnimalOld = dto.TrainingAnimal,
+                TrainingTypeOld = dto.TrainingType,
+            };
+            viewModel.TraineeList = persons
+                .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person, Selected = p.PersonId == dto.PersonId })
                 .ToList();
-
+            viewModel.TrainerList = persons
+               .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person, Selected = p.PersonId == dto.TrainerId })
+               .ToList();
+            viewModel.TrainingTypesList = _staticDropdownService.GetTrainingTypes()
+              .Select(t => new SelectListItem { Value = t.Value, Text = t.Text })
+              .ToList();
+            viewModel.TrainingAnimalList = _staticDropdownService.GetTrainingAnimal()
+               .Select(t => new SelectListItem { Value = t.Value, Text = t.Text })
+               .ToList();
             return View(viewModel);
         }
 
@@ -91,44 +130,67 @@ namespace Apha.BST.Web.Controllers
         {
             if (!ModelState.IsValid)
             {
-                viewModel.Persons = (await _trainingService.GetTraineesAsync())
-                    .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person })
+                var persons = await _trainingService.GetTraineesAsync();
+
+                viewModel.TraineeList = persons
+                    .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person, Selected = p.PersonId == viewModel.TraineeIdOld })
                     .ToList();
+                viewModel.TrainerList = persons
+                   .Select(p => new SelectListItem { Value = p.PersonId.ToString(), Text = p.Person, Selected = p.PersonId == viewModel.TrainerIdOld })
+                   .ToList();
+                viewModel.TrainingTypesList = _staticDropdownService.GetTrainingTypes()
+                  .Select(t => new SelectListItem { Value = t.Value, Text = t.Text, Selected = t.Value == viewModel.TrainingTypeOld })
+                  .ToList();
+                viewModel.TrainingAnimalList = _staticDropdownService.GetTrainingAnimal()
+                   .Select(t => new SelectListItem { Value = t.Value, Text = t.Text, Selected = t.Value == viewModel.TrainingAnimalOld })
+                   .ToList();
+                viewModel.TrainingDateTime = viewModel.TrainingDateTimeOld;
                 return View(viewModel);
             }
 
-            var dto = _mapper.Map<EditTrainingDTO>(viewModel);
-            var message = await _trainingService.UpdateTrainingAsync(dto);
+            var editTraining = _mapper.Map<EditTrainingDto>(viewModel);
+            var message = await _trainingService.UpdateTrainingAsync(editTraining);
             TempData["Message"] = message;
-            return RedirectToAction("ViewTrainer");
+            return RedirectToAction("ViewTraining", new { selectedTraineeId = viewModel.TraineeIdOld });            
         }
 
         [HttpGet]
-        public async Task<IActionResult> ViewTrainer(string selectedTraineeId = siteName)
+        public async Task<IActionResult> ViewTraining(string selectedTraineeId = "")
         {
-            var allPersons = await _personRepository.GetAllAsync(); // PersonID + Person
-            var allTrainees = _mapper.Map<IEnumerable<TraineeViewModel>>(allPersons);
+            var allPersons = await _personService.GetAllPersonAsync();
+            var allTrainee = _mapper.Map<IEnumerable<TraineeViewModel>>(allPersons);
 
-            IEnumerable<TrainingViewModel> filteredTrainings;
+            var viewSelectList = allTrainee
+               .Select(p => new SelectListItem
+               {
+                   Text = p.Person,
+                   Value = p.PersonID.ToString(),
 
-            if (selectedTraineeId == "All")
-            {
-                var allTrainings = await _trainingService.GetAllTrainingsAsync(); // <-- new method to get all
-                filteredTrainings = _mapper.Map<IEnumerable<TrainingViewModel>>(allTrainings);
-            }
-            else
-            {
-                var trainingDtos = await _trainingService.GetTrainingByTraineeAsync(selectedTraineeId);
-                filteredTrainings = _mapper.Map<IEnumerable<TrainingViewModel>>(trainingDtos);
-            }
+               })
+               .ToList();
 
             var model = new TrainingListViewModel
             {
-                AllTrainees = allTrainees,
-                FilteredTrainings = filteredTrainings,
-                SelectedTraineeId = selectedTraineeId
-            };
+                AllTrainees = viewSelectList
 
+            };
+            IEnumerable<TrainingViewModel> filteredTrainings;
+            if (!string.IsNullOrEmpty(selectedTraineeId) && selectedTraineeId != "0")
+            {
+                if (selectedTraineeId == traineeAll)
+                {
+                    var allTrainings = await _trainingService.GetAllTrainingsAsync(); // <-- new method to get all
+                    filteredTrainings = _mapper.Map<IEnumerable<TrainingViewModel>>(allTrainings);
+                }
+                else
+                {
+                    var trainingDtos = await _trainingService.GetTrainingByTraineeAsync(selectedTraineeId);
+                    filteredTrainings = _mapper.Map<IEnumerable<TrainingViewModel>>(trainingDtos);
+                }
+
+                model.FilteredTrainings = filteredTrainings;
+                model.SelectedTraineeId = selectedTraineeId;
+            }
             return View(model);
         }
 
@@ -137,7 +199,7 @@ namespace Apha.BST.Web.Controllers
         [HttpGet]
         public async Task<IActionResult> TrainerHistory(int selectedTrainerId = 0, string selectedSpecies = "Cattle")
         {
-            var allPersons = await _personRepository.GetAllAsync();
+            var allPersons = await _personService.GetAllPersonAsync();
             var allTrainers = _mapper.Map<List<TraineeViewModel>>(allPersons);
 
             var historyDto = await _trainingService.GetTrainerHistoryAsync(selectedTrainerId, selectedSpecies);
@@ -145,14 +207,69 @@ namespace Apha.BST.Web.Controllers
             // FIX: Map DTO to Model
             var historyData = _mapper.Map<List<TrainingHistoryModel>>(historyDto);
 
+
             var model = new TrainerHistoryViewModel
             {
                 SelectedTrainerId = selectedTrainerId,
                 SelectedSpecies = selectedSpecies,
                 AllTrainers = allTrainers,
-                HistoryDetails = historyData,                
+                HistoryDetails = historyData,
+                TrainingAnimalList = _staticDropdownService.GetTrainingAnimal()
             };
+            var trainerSelectList = allTrainers
+                .Select(trainer => new SelectListItem
+                {
+                    Value = trainer.PersonID.ToString(),
+                    Text = trainer.Person,
+                    Selected = trainer.PersonID == selectedTrainerId
+                })
+                .ToList();
 
+            // Insert the "All people" option at the top of the list
+            trainerSelectList.Insert(0, new SelectListItem
+            {
+                Value = "0",
+                Text = "All people",               
+            });
+
+            model.AllTrainersList = trainerSelectList;           
+            return View(model);
+        }
+
+        // For TrainerTrained
+        [HttpGet]
+        public async Task<IActionResult> TrainerTrained(int selectedTrainerId = 0)
+        {
+            var allPersons = await _personService.GetAllPersonAsync();
+            var allTrainers = _mapper.Map<List<TrainerTrainedModel>>(allPersons);
+
+            var trainedList = selectedTrainerId != 0
+                ? await _trainingService.GetTrainerTrainedAsync(selectedTrainerId)
+                : new List<TrainerTrainedDto>();
+
+
+            var model = new TrainerTrainedViewModel
+            {
+                SelectedTrainerId = selectedTrainerId,
+                AllTrainers = allTrainers,
+                TraineeTrainingDetails = trainedList.ToList()
+            };
+            var trainerList = allTrainers
+                .Select(t => new SelectListItem
+                {
+                    Value = t.PersonID.ToString(),
+                    Text = t.Person,
+                    Selected = t.PersonID == selectedTrainerId
+                }).ToList();
+
+            trainerList.Insert(0, new SelectListItem
+            {
+                Value = "0",
+                Text = "Select trainer",
+                Selected = selectedTrainerId == 0
+            });
+
+            model.TrainerSelectList = trainerList;          
 
             return View(model);
         }
@@ -161,9 +278,14 @@ namespace Apha.BST.Web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteTraining(int traineeId, string species, DateTime dateTrained)
         {
+            if(!ModelState.IsValid) 
+            {
+                TempData["Message"] = "Invalid data provided for deletion.";
+                return RedirectToAction(nameof(ViewTraining), new { selectedTraineeId = traineeId });
+            }
             var message = await _trainingService.DeleteTrainingAsync(traineeId, species, dateTrained);
             TempData["Message"] = message;
-            return RedirectToAction(nameof(ViewTrainer), new { selectedTraineeId = traineeId });
+            return RedirectToAction(nameof(ViewTraining), new { selectedTraineeId = traineeId });
         }
